+++ packages/riverpod_generator/lib/src/templates/notifier.dart
@@ -1,4 +1,3 @@
-import 'package:analyzer/dart/element/type.dart';
 import 'package:analyzer_buffer/analyzer_buffer.dart';
 import 'package:riverpod_analyzer_utils/riverpod_analyzer_utils.dart';
 
@@ -63,29 +62,19 @@ abstract class $notifierBaseName$genericsDefinition extends $baseClass {
 
     _writeBuild(buffer);
 
-    final String buildVar;
-    final String buildVarUsage;
-    switch ((
-      provider.providerElement.createdType,
-      provider.providerElement.valueTypeNode,
-    )) {
-      case (SupportedCreatedType.value, VoidType()):
-        buildVar = '';
-        buildVarUsage = 'null';
-      case _:
-        buildVar = 'final created = ';
-        buildVarUsage = 'created';
-    }
+    final buildVarUsage = switch (provider.parameters) {
+      [] => 'build',
+      [_, ...] => '() => build($paramsPassThrough)',
+    };
 
     buffer.writeln('''
   @\$mustCallSuper
   @override
   void runBuild() {
-    ${buildVar}build($paramsPassThrough);
     final ref = this.ref as \$Ref<${provider.providerElement.exposedTypeNode}, ${provider.providerElement.valueTypeNode.toCode()}>;
     final element = ref.element as \$ClassProviderElement<AnyNotifier<${provider.providerElement.exposedTypeNode}, ${provider.providerElement.valueTypeNode.toCode()}>,
           ${provider.providerElement.exposedTypeNode}, Object?, Object?>;
-    element.handleValue(ref, $buildVarUsage);
+    element.handleCreate(ref, $buildVarUsage);
   }
 }
 ''');