+++ packages/riverpod_generator/test/integration/dependencies2.g.dart
@@ -224,7 +224,6 @@ abstract class _$NotifierWithDependencies extends $Notifier<int> {
   @$mustCallSuper
   @override
   void runBuild() {
-    final created = build();
     final ref = this.ref as $Ref<int, int>;
     final element =
         ref.element
@@ -234,7 +233,7 @@ abstract class _$NotifierWithDependencies extends $Notifier<int> {
               Object?,
               Object?
             >;
-    element.handleValue(ref, created);
+    element.handleCreate(ref, build);
   }
 }
 
@@ -343,7 +342,6 @@ abstract class _$NotifierFamilyWithDependencies extends $Notifier<int> {
   @$mustCallSuper
   @override
   void runBuild() {
-    final created = build(id: _$args);
     final ref = this.ref as $Ref<int, int>;
     final element =
         ref.element
@@ -353,7 +351,7 @@ abstract class _$NotifierFamilyWithDependencies extends $Notifier<int> {
               Object?,
               Object?
             >;
-    element.handleValue(ref, created);
+    element.handleCreate(ref, () => build(id: _$args));
   }
 }