+++ packages/riverpod_generator/test/integration/scopes.g.dart
@@ -47,7 +47,6 @@ abstract class _$ScopedClass extends $Notifier<int> {
   @$mustCallSuper
   @override
   void runBuild() {
-    final created = build();
     final ref = this.ref as $Ref<int, int>;
     final element =
         ref.element
@@ -57,7 +56,7 @@ abstract class _$ScopedClass extends $Notifier<int> {
               Object?,
               Object?
             >;
-    element.handleValue(ref, created);
+    element.handleCreate(ref, build);
   }
 }
 
@@ -138,7 +137,6 @@ abstract class _$ScopedClassFamily extends $Notifier<int> {
   @$mustCallSuper
   @override
   void runBuild() {
-    final created = build(_$args);
     final ref = this.ref as $Ref<int, int>;
     final element =
         ref.element
@@ -148,6 +146,6 @@ abstract class _$ScopedClassFamily extends $Notifier<int> {
               Object?,
               Object?
             >;
-    element.handleValue(ref, created);
+    element.handleCreate(ref, () => build(_$args));
   }
 }