+++ website/docs/concepts/about_codegen/provider_type/async_class_future.g.dart
@@ -41,7 +41,6 @@ abstract class _$Example extends $AsyncNotifier<String> {
   @$mustCallSuper
   @override
   void runBuild() {
-    final created = build();
     final ref = this.ref as $Ref<AsyncValue<String>, String>;
     final element =
         ref.element
@@ -51,6 +50,6 @@ abstract class _$Example extends $AsyncNotifier<String> {
               Object?,
               Object?
             >;
-    element.handleValue(ref, created);
+    element.handleCreate(ref, build);
   }
 }