+++ website/docs/concepts2/offline/json_persist.g.dart
@@ -56,7 +56,6 @@ abstract class _$TodoListBase extends $AsyncNotifier<List<Todo>> {
   @$mustCallSuper
   @override
   void runBuild() {
-    final created = build();
     final ref = this.ref as $Ref<AsyncValue<List<Todo>>, List<Todo>>;
     final element =
         ref.element
@@ -66,7 +65,7 @@ abstract class _$TodoListBase extends $AsyncNotifier<List<Todo>> {
               Object?,
               Object?
             >;
-    element.handleValue(ref, created);
+    element.handleCreate(ref, build);
   }
 }