+++ website/docs/concepts2/offline/wait_persist.g.dart
@@ -61,7 +61,6 @@ abstract class _$TodoListBase extends $AsyncNotifier<List<Todo>> {
   @$mustCallSuper
   @override
   void runBuild() {
-    final created = build();
     final ref = this.ref as $Ref<AsyncValue<List<Todo>>, List<Todo>>;
     final element =
         ref.element
@@ -71,7 +70,7 @@ abstract class _$TodoListBase extends $AsyncNotifier<List<Todo>> {
               Object?,
               Object?
             >;
-    element.handleValue(ref, created);
+    element.handleCreate(ref, build);
   }
 }