+++ website/docs/migration/from_change_notifier/migrated/migrated.g.dart
@@ -42,7 +42,6 @@ abstract class _$MyNotifier extends $AsyncNotifier<List<Todo>> {
   @$mustCallSuper
   @override
   void runBuild() {
-    final created = build();
     final ref = this.ref as $Ref<AsyncValue<List<Todo>>, List<Todo>>;
     final element =
         ref.element
@@ -52,6 +51,6 @@ abstract class _$MyNotifier extends $AsyncNotifier<List<Todo>> {
               Object?,
               Object?
             >;
-    element.handleValue(ref, created);
+    element.handleCreate(ref, build);
   }
 }