+++ packages/riverpod/CHANGELOG.md
@@ -1,5 +1,23 @@
 ## Unreleased minor
 
+- Added an alternative way to combine asynchronous providers.
+  This can be done by using `AsyncValue.requireValue` inside `FutureProvider`/`AsyncNotifier`
+  like so:
+
+  ```dart
+  final sumProvider = FutureProvider((ref) { // Not async
+    AsyncValue<int> a = ref.watch(aProvider);
+    AsyncValue<int> b = ref.watch(bProvider);
+
+    // The following is safe if used inside the init function of providers.
+    return a.requireValue + b.requireValue;
+  })
+  ```
+
+  This enables synchronously combining asynchronous providers.
+
+  See [AsyncValue.requireValue](https://pub.dev/documentation/riverpod/latest/riverpod/AsyncValue/requireValue.html)
+
 - Fixed a bug with scoping when specifying `dependencies: [...]`
 - Added `Override.origin`. This enables knowing which provider is associated with an override.
 - Fix a regression with `AsyncLoading.isRefreshing/isReloading`